moved appstore code to separate package to reuse in both main and renderer
