+++ libs/appstore/README.md
@@ -1,0 +1,3 @@
+# appstore-tools
+
+Contains tooling to interact with appstore/portal/devhub.
