+++ libs/appstore/src/appstore/index.ts
@@ -1,0 +1,4 @@
+export * from './appstore-app-client';
+export * from './types';
+export * from './zomes/appstore-zome-client';
+export * from './zomes/portal-zome-client';
