+++ libs/appstore/src/devhub/index.ts
@@ -1,0 +1,5 @@
+export * from './devhub-app-client';
+export * from './types';
+export * from './zomes/apphub-zome-client';
+export * from './zomes/dnahub-zome-client';
+export * from './zomes/zomehub-zome-client';
