+++ libs/appstore/src/mere-memory/index.ts
@@ -1,0 +1,1 @@
+export * from './zomes/mere-memory-zome-client';
